+++ packages/stryker-typescript/src/helpers/tsHelpers.ts
@@ -4,6 +4,7 @@ import { CONFIG_KEY_OPTIONS, CONFIG_KEY_FILE } from './keys';
 import { Config } from 'stryker-api/config';
 import * as ts from 'typescript';
 import * as path from 'path';
+import * as semver from 'semver';
 
 export function createProgram(inputFiles: File[], strykerConfig: Config) {
   const files = inputFiles
@@ -11,10 +12,18 @@ export function createProgram(inputFiles: File[], strykerConfig: Config) {
     .map(file => file.name);
   const options = getTSConfig(strykerConfig);
 
-  return ts.createProgram(files, options);
+  return ts.createProgram(files, options || {});
 }
 
-export function getTSConfig(strykerConfig: Config): ts.CompilerOptions {
+export function parseFile(file: TextFile, target: ts.ScriptTarget | undefined) {
+  return ts.createSourceFile(
+    file.name,
+    file.content,
+    target || ts.ScriptTarget.ES5,
+    /*setParentNodes*/ true);
+}
+
+export function getTSConfig(strykerConfig: Config): ts.CompilerOptions | undefined {
   return strykerConfig[CONFIG_KEY_OPTIONS];
 }
 
@@ -34,6 +43,15 @@ export function getProjectDirectory(config: Config) {
   return path.dirname(config[CONFIG_KEY_FILE] || '.');
 }
 
+/**
+ * Verifies that the installed version of typescript satisfies '>=2.5` and otherwise: throws an exception
+ */
+export function guardTypescriptVersion() {
+  if (!semver.satisfies(ts.version, '>=2.5')) {
+    throw new Error(`Installed typescript version ${ts.version} is not supported by stryker-typescript. Please install version 2.5 or higher (\`npm install typescript@^2.5\`).`);
+  }
+}
+
 const printer = ts.createPrinter({
   removeComments: false,
   newLine: os.EOL === '\r\n' ? ts.NewLineKind.CarriageReturnLineFeed : ts.NewLineKind.LineFeed
@@ -44,17 +62,28 @@ export function printNode(node: ts.Node, originalSourceFile: ts.SourceFile): str
 }
 
 const allExtensions: string[] = Object.keys(ts.Extension).map(extension => ts.Extension[extension as any]);
-export function isToBeTranspiled(file: File) {
+export function isTypescriptFile(file: File) {
   return file.kind === FileKind.Text &&
-    file.transpiled &&
-    allExtensions.some(extension => file.name.endsWith(extension)) && !file.name.endsWith('.d.ts');
+    allExtensions.some(extension => file.name.endsWith(extension));
 }
 
-
-export function filterOutTypescriptFiles(files: File[]): TextFile[] {
-  return files.filter(isToBeTranspiled) as TextFile[];
+/**
+ * Determines whether or not given file is a typescript header file (*.d.ts)
+ */
+export function isHeaderFile(file: File) {
+  return file.name.endsWith('.d.ts');
 }
 
-export function filterEmpty<T>(input: (T | undefined | null)[]): T[] {
+/**
+ * Returns all the files that are considered typescript files (text files with *.ts or something like that)
+ */
+export function filterTypescriptFiles(files: File[]): TextFile[] {
+  return files.filter(isTypescriptFile) as TextFile[];
+}
+
+/**
+ * Returns all items that are NOT undefined or null
+ */
+export function filterNotEmpty<T>(input: (T | undefined | null)[]): T[] {
   return input.filter(item => item !== void 0 && item !== null) as T[];
 }