+++ packages/stryker-typescript/testResources/binaryExpressionTestFiles/testfileWith1ExclamationEqualsEqualsTokenMutation.ts
@@ -1,1 +1,0 @@
-let exclamationequalsequals = false !== false;