+++ packages/stryker-typescript/testResources/binaryExpressionTestFiles/testfileWith1ExclamationEqualsTokenMutation.ts
@@ -1,1 +1,0 @@
-let exclamationequals = false != false;