+++ packages/stryker-typescript/testResources/binaryExpressionTestFiles/testfileWith1PercentageTokenMutation.ts
@@ -1,1 +1,0 @@
-let modulo = 2 % 1;