+++ packages/stryker-karma-runner/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.11.1"></a>
+## [0.11.1](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.11.0...stryker-karma-runner@0.11.1) (2017-11-14)
+
+
+### Bug Fixes
+
+* **init stryker typescript:** Add init section ([#459](https://github.io/stryker-mutator/stryker/issues/459)) ([c4510d5](https://github.io/stryker-mutator/stryker/commit/c4510d5)), closes [#454](https://github.io/stryker-mutator/stryker/issues/454)
+
+
+
+
+<a name="0.11.0"></a>
+# [0.11.0](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.10.0...stryker-karma-runner@0.11.0) (2017-11-13)
+
+
+### Features
+
+* **karma:** Ability to override options ([#448](https://github.io/stryker-mutator/stryker/issues/448)) ([ee5d008](https://github.io/stryker-mutator/stryker/commit/ee5d008)), closes [#434](https://github.io/stryker-mutator/stryker/issues/434)
+
+
+
+
+<a name="0.10.0"></a>
+# [0.10.0](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.9.1...stryker-karma-runner@0.10.0) (2017-11-13)
+
+
+### Features
+
+* **mocha 4:** Add support for mocha version 4 ([#455](https://github.io/stryker-mutator/stryker/issues/455)) ([de6ae4f](https://github.io/stryker-mutator/stryker/commit/de6ae4f))
+
+
+
+
 <a name="0.9.1"></a>
 ## [0.9.1](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.9.0...stryker-karma-runner@0.9.1) (2017-10-24)